// Frontend/src/app/models/user.model.ts
export interface NguoiDung {
    maND: number;
    taiKhoan: string;
    matKhau: string
    email: string
    hoTen: string
    ngaySinh: string
    gioiTinh: string
    anhDaiDien: string
    diaChi: string
    sdt: string
    ngayThamGia: string
    trangThai: number;
    maQuyen: number;

    tenQuyen: string;
}